feat: move `int` allocator things to separate folder
