+++ packages/compiler-pico-c/src/arch/x86/backend/reg-allocator/index.ts
@@ -1,2 +1,1 @@
-export * from './X86BasicRegAllocator';
-export * from './X86RegOwnershipTracker';
+export * from './int';
