+++ packages/compiler-pico-c/src/arch/x86/backend/reg-allocator/int/index.ts
@@ -1,0 +1,2 @@
+export * from './X86BasicRegAllocator';
+export * from './X86RegOwnershipTracker';
