+++ packages/compiler-pico-c/src/arch/x86/constants/regs.ts
@@ -1,9 +1,14 @@
+import type {
+  X86BitsMode,
+  X86IntBitsMode,
+  X86RegName,
+} from '@ts-c-compiler/x86-assembler';
+
 import { CCompilerArch } from '#constants';
-import { X86IntBitsMode, X86RegName } from '@ts-c-compiler/x86-assembler';
 import { COMPILER_REGISTERS_SET } from '@ts-c-compiler/x86-assembler';
 import {
-  X87StackRegName,
   X87_STACK_REGISTERS,
+  type X87StackRegName,
 } from '@ts-c-compiler/x86-assembler';
 
 export type X86RegsParts = {
@@ -19,6 +24,11 @@ export type X86IntRegTree = {
   unavailable?: boolean;
 };
 
+export type X87FloatReg = {
+  name: X87StackRegName;
+  size: X86BitsMode;
+};
+
 export type RegsMap = {
   stack: X86RegName;
   addressing: Array<X86RegName>;
@@ -29,7 +39,7 @@ export type RegsMap = {
   };
   int: X86IntRegTree[];
   float: {
-    x87: Readonly<X87StackRegName[]>;
+    x87: Readonly<X87FloatReg[]>;
   };
 };
 
@@ -130,7 +140,10 @@ export const createX86RegsMap = (): Record<CCompilerArch, RegsMap> => ({
       },
     ],
     float: {
-      x87: X87_STACK_REGISTERS,
+      x87: X87_STACK_REGISTERS.map(stackRegName => ({
+        size: 0xa,
+        name: stackRegName,
+      })),
     },
   },
 });
